handle initialize exceptions

+++ src/main/java/com/klindziuk/mob/browser/BrowserDriver.java
@@ -5,15 +5,20 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 
+import io.netty.util.internal.logging.InternalLogger;
+import io.netty.util.internal.logging.InternalLoggerFactory;
 import net.lightbody.bmp.BrowserMobProxy;
 
 import java.io.IOException;
 
 public final class BrowserDriver {
+	private final static InternalLogger logger = InternalLoggerFactory.getInstance("BrowserDriver");
+	private static final String INVALID_BROWSER_NAME_MESSAGE = "Cannot initialize WebDriver without valid value  of browser name.";
 	private static final String CHROME = "chrome";
 	private static final String FIREFOX = "firefox";
 	private BrowserMobProxy proxy;
 	private BrowserInitializer initializer;
+	private DesiredCapabilities capabilities;
 
 	public BrowserDriver(BrowserMobProxy proxy, BrowserInitializer initializer) {
 		this.proxy = proxy;
@@ -21,31 +26,47 @@ public final class BrowserDriver {
 	}
 
 	public WebDriver selectDriver(String browserName) {
+		if (null == browserName || browserName.isEmpty()) {
+			logger.error(INVALID_BROWSER_NAME_MESSAGE);
+			return null;
+		}
 		switch (browserName) {
 		case CHROME: {
-			DesiredCapabilities capabilities = initializer.getProxyChromeCapabilities(proxy);
+			try {
+				capabilities = initializer.getProxyChromeCapabilities(proxy);
+			} catch (IllegalArgumentException iaex) {
+				logger.error(iaex.getMessage(), iaex);
+			}
 			return new ChromeDriver(capabilities);
 		}
 		case FIREFOX: {
-			DesiredCapabilities capabilities = initializer.getProxyFFCapabilities(proxy);
+			try {
+				capabilities = initializer.getProxyFFCapabilities(proxy);
+			} catch (IllegalArgumentException iaex) {
+				logger.error(iaex.getMessage(), iaex);
+			}
 			return new FirefoxDriver(capabilities);
 		}
 		default: {
+			logger.error(INVALID_BROWSER_NAME_MESSAGE);
 			return null;
 		}
 		}
 	}
-	
+
 	public void killDriver(String browserName) {
+		if (null == browserName || browserName.isEmpty()) {
+			return;
+		}
 		try {
 			switch (browserName) {
-			case CHROME : {
+			case CHROME: {
 				Runtime.getRuntime().exec("taskkill /f /IM chromedriver.exe");
 			}
-			case FIREFOX : {
+			case FIREFOX: {
 				Runtime.getRuntime().exec("taskkill /f /IM geckodriver.exe");
 			}
-		 }
+			}
 		} catch (IOException ioex) {
 			ioex.printStackTrace();
 		}
